long_value is stored as a string in mongodb and recommender expects a Long type

+++ service/src/main/java/com/hypemrecommender/dal/UserDao.java
@@ -7,5 +7,5 @@ package com.hypemrecommender.dal;
  * Time: 22:01
  */
 public interface UserDao {
-    int getUserId(String username);
+    long getUserId(String username);
 }
